* klist.c (one_addr): gethostbyaddr takes an int as second argument
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12798 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/clients/klist/klist.c
@@ -670,7 +670,7 @@ void one_addr(a)
 		freehostent(h);
 	    }
 #else
-	    h = gethostbyaddr(a->contents, a->length, af);
+	    h = gethostbyaddr(a->contents, (int) a->length, af);
 	    if (h) {
 		printf("%s", h->h_name);
 	    }
